+++ a1/controller.py
@@ -30,23 +30,31 @@ mdc_logger = Logger(name=__name__)
 
 def _try_func_return(func):
     """
-    generic caller that returns the apporp http response if exceptions are raised
+    helper method that runs the function and returns a detailed http response if an exception is raised.
     """
     try:
         return func()
-    except (ValidationError, exceptions.PolicyTypeAlreadyExists, exceptions.CantDeleteNonEmptyType):
-        return "", 400
-    except (exceptions.PolicyTypeNotFound, exceptions.PolicyInstanceNotFound):
-        return "", 404
-    except (RejectedByBackend, NotConnected, BackendError):
+    except (ValidationError, exceptions.PolicyTypeAlreadyExists, exceptions.CantDeleteNonEmptyType) as exc:
+        msg = repr(exc)
+        mdc_logger.warning("Request failed, returning 400: {0}".format(msg))
+        return msg, 400
+    except (exceptions.PolicyTypeNotFound, exceptions.PolicyInstanceNotFound) as exc:
+        msg = repr(exc)
+        mdc_logger.warning("Request failed, returning 404: {0}".format(msg))
+        return msg, 404
+    except (RejectedByBackend, NotConnected, BackendError) as exc:
         """
-        These are SDL errors. At the time of development here, we do not have a good understanding which of these errors are "try again later it may work"
-        and which are "never going to work". There is some discussion that RejectedByBackend is in the latter category, suggesting it should map to 400,
-        but until we understand the root cause of these errors, it's confusing to clients to give them a 400 (a "your fault" code) because they won't know how to fix
+        These are SDL errors. At the time of development here, we do not have a good understanding
+        which of these errors are "try again later it may work" and which are "never going to work".
+        There is some discussion that RejectedByBackend is in the latter category, suggesting it
+        should map to 400, but until we understand the root cause of these errors, it's confusing
+        to clients to give them a 400 (a "your fault" code) because they won't know how to fix.
         For now, we log, and 503, and investigate the logs later to improve the handling/reporting.
         """
         # mdc_logger.exception(exc)  # waiting for https://jira.o-ran-sc.org/browse/RIC-39
-        return "", 503
+        msg = repr(exc)
+        mdc_logger.warning("Request failed, returning 503: {0}".format(msg))
+        return msg, 503
 
     # let other types of unexpected exceptions blow up and log